Refactoring how the timetable datastructure works
Instead of just storing which employees take up which shifts,
it ended up being a bit more straight forward to just have timetable
represent full days of the week, which are then assigned employees.

This allows us to micro assign people over certain contiguous blocks
and represent the schedule in a fashion that is easy to iterate over
and display.